Lexer enhancement, replace string_view with string
The lexer stores now in addition individual parts of parsed versions for
easier access and prevention of the necessity of reparsing the parts.

The usage of string_view has been replaced with string. The handling is
easier and SSO is to be expected to be the normal case, because
individual parts of semver strings are very small. It is not to be
excepted to exeed the usual limits (GCC=23, Clang=15).